RDMA: Use PRIu64 to print a uint64_t
In /usr/include/infiniband/verbs.h:
struct ibv_wc {
	uint64_t                wr_id;
	enum ibv_wc_status      status;
	[...]

Fix this cppcheck warning:
pcap-rdmasniff.c:162:4: warning: %lld in format string (no. 1) requires
'long long' but the argument type is 'unsigned long long'.
[invalidPrintfArgType_sint]
   fprintf(stderr, "failed WC wr_id %lld status %d/%s\n",
   ^

Remove an unneeded cast.

+++ pcap-rdmasniff.c
@@ -159,8 +159,8 @@ rdmasniff_read(pcap_t *handle, int max_packets, pcap_handler callback, u_char *u
 		}
 
 		if (wc.status != IBV_WC_SUCCESS) {
-			fprintf(stderr, "failed WC wr_id %lld status %d/%s\n",
-				(unsigned long long) wc.wr_id,
+			fprintf(stderr, "failed WC wr_id %" PRIu64 " status %d/%s\n",
+				wc.wr_id,
 				wc.status, ibv_wc_status_str(wc.status));
 			continue;
 		}
